Source location reporting and macros __LINE__, __FILE__, __DATE__ and __TIME__.

+++ src/lib/input.h
@@ -9,8 +9,8 @@ struct sp_input {
   struct sp_input *next;
   uint16_t file_id;
   int base_cond_level;
-  int size;
-  int pos;
+  size_t size;
+  size_t pos;
   unsigned char data[];
 };
